(#283) Extract parser related stuff to a separate trans unit
